requirements.txt
pyramid
pyramid_chameleon
pyramid_tm
SQLAlchemy
transaction
zope.sqlalchemy
waitress
